Reverted publication of AbstractContextualStorageManager, extracted CdiInstantiator implementation into AbstractCdiInstantiator
